% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcsinhLink.R
\name{arcsinhLink}
\alias{arcsinhLink}
\title{arcsinhLink
arcsinh link function for GLM}
\usage{
arcsinhLink()
}
\value{
An object of class "link-glm"
}
\description{
arcsinhLink
arcsinh link function for GLM
}
