% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcsinhTransform.R
\name{arcsinhTransform}
\alias{arcsinhTransform}
\title{arcsinhTransform}
\usage{
arcsinhTransform(counts, design = NULL, lib.size = NULL,
  normalize.method = "none", span = 0.5, plot = FALSE,
  save.plot = FALSE, ...)
}
\arguments{
\item{counts}{\code{DGEList} as for \code{voom}.}

\item{design}{model matrix created by \code{computeStat()} as for \code{voom}.}

\item{lib.size}{library size created by \code{computeStat()} as for \code{voom}.}

\item{normalize.method}{no need to use this argument because we do not use logCPM values}

\item{span}{argument as for \code{voom}.}

\item{plot}{logical, as for \code{voom}.}

\item{save.plot}{logical, as for \code{voom}.}
}
\value{
An EList object as \code{voom}. E is numeric matrix of transformed counts on the arcsine scale
}
\description{
This is a utility function used by the higher level function \code{computeStat} to compute arcsine transformation and weights for the glm fit.
}
