% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asinh_voom.R
\name{asinh_voom}
\alias{asinh_voom}
\title{Modifications to \code{\link[limma]{voom}}}
\usage{
asinh_voom(counts, design = NULL, sj, span = 0.5, plot = FALSE)
}
\arguments{
\item{counts}{A numeric matrix of otu_table.}

\item{design}{A design matrix with rows corresponding to samples and columns to coefficients to be estimated. Defaults to the unit vector meaning that samples are treated as replicates.}

\item{sj}{A numeric vector containing library size normalization factor based on the median-ratio method.}

\item{span}{A numeric. Width of the lowess smoothing window as a proportion.}

\item{plot}{A logical, should a plot of the mean-variance trend be displayed?}
}
\value{
A matrix of variance stablized counts.
}
\description{
Computes the observation level weight based on the mean-variance relationship.
}
