% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboot_indices.R
\name{bboot_indices}
\alias{bboot_indices}
\title{bboot_indices}
\usage{
bboot_indices(x, b, time, k, blks_first_index)
}
\arguments{
\item{x}{a data frame. It must have \code{Index} column that will be added by \code{bboot_phyloseq}.}

\item{b}{a numeric. Block size.}

\item{time}{a character. Name of the time variable.}

\item{k}{a numeric. Number of overlapping blocks per subject that will be added by \code{bboot_phyloseq}.}

\item{blks_first_index}{a vector of positive integer. This will be added by \code{bboot_phyloseq} (same for all subjects.)}
}
\value{
A list of indices to include in the block bootstrap realization.
}
\description{
This function is called within \code{bboot_phyloseq} for each subject's data frame.
}
