% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboot_optblocksize.R
\name{bboot_optblocksize}
\alias{bboot_optblocksize}
\title{bboot_optblocksize}
\usage{
bboot_optblocksize(ps, R, RR, factors, time, subjectidvar = "SubjectID", lI,
  omega = 0.6)
}
\arguments{
\item{ps}{Observed \code{phyloseq} class object.}

\item{R}{Number of block bootstrap realization.}

\item{RR}{Number of double block bootstrap realization.}

\item{factors}{vector of factor variable(s) in the sample data of ps.}

\item{time}{Time variable at repeated observations.}

\item{subjectidvar}{variable name for SubjectID}

\item{lI}{initial block size}

\item{omega}{percentage of repeated observations for subsampling}
}
\value{
list of ``MSE`` in calculating K with block sizes lC=1:(lI-1), ``Khat`` values with block sizes lC=1:(lI-1) and ``Khat.obs`` with lI
}
\description{
Finding optimal block size
}
