% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboot_phyloseq.R
\name{bboot_phyloseq}
\alias{bboot_phyloseq}
\title{bboot_phyloseq}
\usage{
bboot_phyloseq(ps, b, time)
}
\arguments{
\item{ps}{Observed \code{phyloseq} class object.}

\item{b}{Block lenght to account for dependence within-subject.}

\item{time}{Time variable at repeated measures.}
}
\value{
Block bootstrap phyloseq object as a list.
}
\description{
Creates block bootstrap phyloseq object.
}
