% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLongIndices.R
\name{bootLongIndices}
\alias{bootLongIndices}
\title{bootLongIndices}
\usage{
bootLongIndices(x, time_var, b, L, blks.first.index)
}
\arguments{
\item{x}{A dataframe. It must have \code{Index} column that will be added by \code{bootLongPhyloseq}.}

\item{time_var}{Character string. The name of the variable to map time in the plot (integer vector).}

\item{b}{A numeric. The block size to account for the dependence within-subject.}

\item{L}{A numeric. Number of overlapping blocks per subject that will be added by \code{bootLongPhyloseq}.}

\item{blks.first.index}{A vector of positive integer. This will be added by \code{bootLongPhyloseq} (same for all subjects.)}
}
\value{
A list of indices to include in the block bootstrap realization.
}
\description{
This function is called within \code{bootLongPhyloseq} for
each subject's data frame.
}
