% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLongMSEPsi.R
\name{bootLongMSEPsi}
\alias{bootLongMSEPsi}
\title{bootLongMSEPsi}
\usage{
bootLongMSEPsi(ps, main_factor, time_var, subjectID_var, b, R, RR, qj, Wj,
  Khat.obs = NULL, T.obs.full = NULL)
}
\arguments{
\item{ps}{phyloseq object}

\item{main_factor}{Character string. The name of the variable to map main factor in the plot (factor vector).}

\item{time_var}{Character string. The name of the variable to map time in the plot (integer vector).}

\item{subjectID_var}{Character string. The name of the variable to map Subject Ids in the plot (factor vector).}

\item{b}{A numeric. The block size to account for the dependence within-subject.}

\item{R}{A numeric. The number of block bootstrap realizations.}

\item{RR}{A numeric. The number of double block bootstrap realizations.}

\item{qj}{number of repeated observations for j-th subject}

\item{Wj}{subset of repeated observations for j-th subject}

\item{Khat.obs}{second element of the output of \code{bootLongPsi} evaluated using the initial block length and full data}

\item{T.obs.full}{A numeric vector. The numeric vector of observed statistic.}
}
\value{
list of MSE computed with block length b, Khat is all subsamples, Khat with initial block length
}
\description{
Compute MSE in computing K = two-sided probability with different block sizes
}
