% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLongPsi.R
\name{bootLongPsi}
\alias{bootLongPsi}
\title{bootLongPsi}
\usage{
bootLongPsi(ps, main_factor, time_var, subjectID_var, b, R, RR,
  T.obs.full = NULL, ncores)
}
\arguments{
\item{ps}{phyloseq object}

\item{main_factor}{Character string. The name of the variable to map main factor in the plot (factor vector).}

\item{time_var}{Character string. The name of the variable to map time in the plot (integer vector).}

\item{subjectID_var}{Character string. The name of the variable to map Subject Ids in the plot (factor vector).}

\item{b}{A numeric. The block size to account for the dependence within-subject.}

\item{R}{A numeric. The number of block bootstrap realizations.}

\item{RR}{A numeric. The number of double block bootstrap realizations.}

\item{T.obs.full}{A numeric vector. The numeric vector of observed statistic.}

\item{ncores}{A numeric. The number of cores to use in \code{\link[parallel]{mclapply}}.}
}
\value{
a list with first element ``K.val`` - two-sided significance probability.
        second element ``observed statistic``
}
\description{
Compute \eqn{\psi} = two-sided probability with a given block size.
}
