% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLongWorkingCor.R
\name{bootLongWorkingCor}
\alias{bootLongWorkingCor}
\title{compute working correlation using block bootstrap}
\usage{
bootLongWorkingCor(x, b)
}
\arguments{
\item{x}{A numeric vector.}

\item{b}{A numeric. The block size to account for the dependence within-subject.}
}
\value{
A numeric matrix. The working autocorrelation for GEE using block bootstrap method.
}
\description{
compute working correlation using block bootstrap
}
