% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stat_asinh_voom_limma.R
\name{compute_stat}
\alias{compute_stat}
\title{compute_stat}
\usage{
compute_stat(ps, factors)
}
\arguments{
\item{ps}{Observed \code{phyloseq} class object.}

\item{factors}{vector of factor variable(s) in the sample data of ps.}
}
\value{
dataframe with the last column corresponds to taxa names
and other columns are for moderated coefficients for each factors, including Intercept.
}
\description{
This function computes size factors for each samples using \code{estimateSizeFactors}.
Then, it transforms count data to inverse hyperbolic sine transformation.
Using the transformed counts, this function estimates the mean-variance relationship and
appropriate observation-level weights. These weights and the transformed data is used to fit a linear model
for each taxa. The empirical Bayes is applied to compute the shirnkage estimates for each factor.
}
