% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeK.R
\name{ComputeK}
\alias{ComputeK}
\title{ComputeK}
\usage{
ComputeK(ps, b, R, RR, factors, time, T.obs.full = NULL)
}
\arguments{
\item{ps}{Observed \code{phyloseq} class object.}

\item{b}{numeric, block size to account for dependence within-subject.}

\item{R}{Number of block bootstrap realization.}

\item{RR}{Number of double block bootstrap realization.}

\item{factors}{vector of factor variable(s) in the sample data of ps.}

\item{time}{Time variable at repeated observations.}

\item{T.obs.full}{If observed statistic is already computed}
}
\value{
a list with first element ``K.val`` - two-sided significance probability.
        second element ``observed statistic``
}
\description{
Compute K=two-sided probability with given block size
}
