% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeMSE.R
\name{ComputeMSE}
\alias{ComputeMSE}
\title{ComputeMSE}
\usage{
ComputeMSE(ps, qj, Wj, b, R, RR, factors, time, Khat.obs = NULL,
  T.obs.full = NULL)
}
\arguments{
\item{ps}{Observed \code{phyloseq} class object.}

\item{qj}{number of repeated observations for j-th subject}

\item{Wj}{subset of repeated observations for j-th subject}

\item{b}{numeric, block lenght to account for dependence within-subject.}

\item{R}{Number of block bootstrap realization.}

\item{RR}{Number of double block bootstrap realization.}

\item{factors}{vector of factor variable(s) in the sample data of ps.}

\item{time}{Time variable at repeated observations.}

\item{Khat.obs}{second element of the output of \code{ComputeK} evaluated using the initial block length and full data}

\item{T.obs.full}{If observed statistic is already computed}

\item{compute_stat}{}
}
\value{
list of MSE computed with block length b, Khat is all subsamples, Khat with initial block length
}
\description{
Compute MSE in computing K = two-sided probability with different block sizes
}
