% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStat.R
\name{computeStat}
\alias{computeStat}
\title{computeStat}
\usage{
computeStat(ps, main_factor, time_var, subjectID_var, b)
}
\arguments{
\item{ps}{phyloseq object}

\item{main_factor}{Character string. The name of the variable to map main factor in the plot (factor vector).}

\item{time_var}{Character string. The name of the variable to map time in the plot (integer vector).}

\item{subjectID_var}{Character string. The name of the variable to map Subject Ids in the plot (factor vector).}

\item{b}{A numeric. The block size to account for the dependence within-subject.}
}
\value{
dataframe with the last column corresponds to taxa names
and other columns are for estimated coefficients for each covariate using GEE, including Intercept.
}
\description{
Computes the regression coefficients.
}
\details{
Computes the library normalization factors using the median-ratio method, computes the observation-level weight using the mean-variance relationship (modification to \code{\link[limma]{voom}}), computes the residuals using glm \code{\link[MASS]{glm.nb}}, computes the working correlation based on the given block size, computes the regression coefficient using the generalized estimating equation \code{\link[geeM]{geem}}.
}
