% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMic_EDA2.R
\name{LMic_correloram}
\alias{LMic_correloram}
\title{LMic_correloram}
\usage{
LMic_correloram(ps, factors, time, taxon, taxlevel = "Species")
}
\arguments{
\item{ps}{\code{phyloseq} object}

\item{factors}{factor variable in the sample data of ps.}

\item{time}{character, time variable at repeated observations.}

\item{taxon}{numeric, index of taxon to get the variogram}

\item{taxlevel}{character, taxonomy level to put the title}
}
\value{
\code{ggplot2} object of correlogram for one taxon according to the taxon index.
}
\description{
Autocorrelation plot
}
