% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMic_EDA2.R
\name{LMic_correloram_multaxa}
\alias{LMic_correloram_multaxa}
\title{LMic_correloram_multaxa}
\usage{
LMic_correloram_multaxa(ps, factors, time, starttaxa = 1, endtaxa = 4,
  taxlevel = "Species")
}
\arguments{
\item{ps}{\code{phyloseq} object}

\item{factors}{factor, variable in the sample data of ps.}

\item{time}{character, time variable at repeated observations.}

\item{starttaxa}{numeric, first index of taxa ordered by total reads}

\item{endtaxa}{numeric, the last index of taxa ordered by total reads}

\item{taxlevel}{character, taxonomy level to put the title}
}
\value{
\code{ggplot2} object of correlogram for multiple taxa.
}
\description{
LMic_correloram_multaxa
}
