% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMic_EDA2.R
\name{LMic_lagged}
\alias{LMic_lagged}
\title{LMic_lagged}
\usage{
LMic_lagged(ps, factors, time, taxon, lags, taxlevel = "Species")
}
\arguments{
\item{ps}{ps \code{phyloseq} object}

\item{factors}{factor, variable in the sample data of ps.}

\item{time}{character, time variable at repeated observations.}

\item{taxon}{numeric, index of taxon to get the variogram}

\item{lags}{numeric (vector) how many lags to plot}

\item{taxlevel}{character, taxonomy level to put the title}
}
\value{
\code{ggplot2} object of lagged plot for speciec taxon.
}
\description{
LMic_lagged
}
