% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMic_EDA.R
\name{LMic_vario_multaxa}
\alias{LMic_vario_multaxa}
\title{LMic_vario_multaxa}
\usage{
LMic_vario_multaxa(ps, factors, time, starttaxa = 1, endtaxa = 4,
  point = FALSE, taxlevel = "Species")
}
\arguments{
\item{ps}{\code{phyloseq} object}

\item{factors}{factor, variable in the sample data of ps.}

\item{time}{character, time variable at repeated observations.}

\item{starttaxa}{numeric, starting index of taxon: taxa will be ordered inside the function}

\item{endtaxa}{numeric, last index of taxon: taxa will be ordered inside the function}

\item{taxlevel}{character, taxonomy level to put the title}
}
\value{
\code{ggplot} object of variogram for starttaxa:endtaxa taxa
}
\description{
LMic_vario_multaxa
}
