% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longLagPlot.R
\name{longLagPlot}
\alias{longLagPlot}
\title{longLagPlot}
\usage{
longLagPlot(ps, factors, time, taxon, lags, taxlevel = "Species")
}
\arguments{
\item{ps}{ps \code{phyloseq} object}

\item{factors}{factor, variable in the sample data of ps.}

\item{time}{character, time variable at repeated observations.}

\item{taxon}{numeric, index of taxon to get the variogram}

\item{lags}{numeric (vector) how many lags to plot}

\item{taxlevel}{character, taxonomy level to put the title}
}
\value{
\code{ggplot2} object of lagged plot for speciec taxon.
}
\description{
This produce the lagged plot given the taxon index.
}
