% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longPACFSingle.R
\name{longPACFSingle}
\alias{longPACFSingle}
\title{Plots the PACF (Partial Autocorrelation Function) given the taxon index.}
\usage{
longPACFSingle(ps, main_factor, time_var, taxon, taxlevel = "Species")
}
\arguments{
\item{ps}{phyloseq object.}

\item{main_factor}{Character string. The name of the variable to map main factor in the plot (factor vector).}

\item{time_var}{Character string. The name of the variable to map time in the plot (integer vector).}

\item{taxon}{Numeric. The index of taxon to get the correlogram.}

\item{taxlevel}{Character string. The taxonomy level for the plot title.}
}
\value{
\code{ggplot2} object of PACF for one taxon according to the taxon index.
}
\description{
Plots the PACF (Partial Autocorrelation Function) given the taxon index.
}
