% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongVarioMultiple.R
\name{longVarioMultiple}
\alias{longVarioMultiple}
\title{longVarioMultiple}
\usage{
longVarioMultiple(pstr, psres, factors, time, starttaxa = 1,
  endtaxa = 4, point = FALSE, taxlevel = "Species")
}
\arguments{
\item{factors}{factor, variable in the sample data of ps.}

\item{time}{character, time variable at repeated observations.}

\item{starttaxa}{numeric, starting index of taxon: taxa will be ordered inside the function}

\item{endtaxa}{numeric, last index of taxon: taxa will be ordered inside the function}

\item{taxlevel}{character, taxonomy level to make the title}

\item{ps}{\code{phyloseq} object}
}
\value{
\code{ggplot} object of variogram for starttaxa:endtaxa taxa
}
\description{
This function plots the variogram for multiple taxa given the indeices.
}
