% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longVarioSingle.R
\name{longVarioSingle}
\alias{longVarioSingle}
\title{Plot the variogram given the taxon index.}
\usage{
longVarioSingle(ps, main_factor, time_var, subjectID_var, taxon,
  point = FALSE, taxlevel = "Species")
}
\arguments{
\item{ps}{phyloseq object}

\item{main_factor}{Character string. The name of the variable to map main factor in the plot (factor vector).}

\item{time_var}{Character string. The name of the variable to map time in the plot (integer vector).}

\item{subjectID_var}{Character string. The name of the variable to map Subject Ids in the plot (factor vector).}

\item{taxon}{Numeric, index of taxon to get the correlogram}

\item{point}{Logical. Whether variogram with observed values.}

\item{taxlevel}{Character string. The taxonomy level for the plot title}
}
\value{
\code{ggplot} object of variogram
}
\description{
Plot the variogram given the taxon index.
}
