% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sampling_schedule.R
\name{plot_sampling_schedule}
\alias{plot_sampling_schedule}
\title{plot_sampling_schedule}
\usage{
plot_sampling_schedule(ps, time_var, subjectID_var, main_factor,
  theme_manual = set.theme)
}
\arguments{
\item{ps}{phyloseq object}

\item{time_var}{Character string. The name of the variable to map time in the plot (integer vector).}

\item{subjectID_var}{Character string. The name of the variable to map Subject Ids in the plot (factor vector).}

\item{main_factor}{Character string. The name of the variable to map main factor in the plot (factor vector).}

\item{theme_manual}{Character string. Specifying the object of \link[ggplot2]{theme_update}.}
}
\value{
an object of ggplot that displays the sampling schedule.
}
\description{
This function plots the sampling schedule.
}
