% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMic_EDA.R
\name{ps_trans}
\alias{ps_trans}
\title{ps_trans}
\usage{
ps_trans(ps, factors)
}
\arguments{
\item{ps}{\code{phyloseq} object}

\item{factors}{factor, variable in the sample data of ps.}
}
\value{
\code{phyloseq} object with transformed \code{otu_table}
}
\description{
ps_trans
}
