% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssim.R
\docType{data}
\name{pssim}
\alias{pssim}
\title{Simulated \code{'phyloseq'} object.}
\format{An object of class \code{'phyloseq'}.
\describe{
  \item{SubjectID}{IDs of individuals}
  \item{Group}{A factor variable with two levels Case or Control}
  \item{SampleID}{Sample ID for each repeated biological sample}
  \item{Time}{Time at repeated biological sample}
}}
\source{
Simulated based on the pregnancy data. Paper on \href{http://www.pnas.org/content/112/35/11060.short}{where}
}
\usage{
pssim
}
\description{
A \code{'phyloseq'} class object with 50 taxa from 300 biological samples.
}
\examples{
\dontrun{sample_data(pssim)
otu_table(pssim)}
}
\references{
Pratheepa Jeganathan (Simulated dataset)
}
\keyword{datasets}
