% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psTransform.R
\name{psTransform}
\alias{psTransform}
\title{psTransform}
\usage{
psTransform(ps, span = 0.5)
}
\arguments{
\item{ps}{\code{phyloseq} object}

\item{span}{the smoother span. see \code{lowess}}
}
\value{
\code{phyloseq} object with transformed \code{otu_table}
}
\description{
This function ctakes the \code{phyloseq} object, apply
the arcsinh transformation, computes the mean-variance relationship for arcsinh transformed counts, assign a weight for each observation based on the predicted variance.
}
