% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.lag.correlation.R
\name{sample.lag.correlation}
\alias{sample.lag.correlation}
\title{Sample lag autocorrelation}
\usage{
sample.lag.correlation(x, b = 3)
}
\arguments{
\item{x}{A numeric vector.}

\item{b}{A numeric scalar. The block size.}
}
\value{
A matrix of autocorrelation.
}
\description{
Sample lag autocorrelation
}
