% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voom_arcsine.R
\name{voom_arcsine}
\alias{voom_arcsine}
\title{voom_arcsin}
\usage{
voom_arcsine(counts, design = NULL, lib.size = NULL,
  normalize.method = "none", span = 0.5, plot = FALSE,
  save.plot = FALSE, ...)
}
\arguments{
\item{counts}{\code{DGEList} as for \code{voom}.}

\item{design}{model matrix created by \code{compute_stat()} as for \code{voom}.}

\item{lib.size}{library size created by \code{compute_stat()} as for \code{voom}.}

\item{normalize.method}{no need to use this argument because we do not use logCPM values}

\item{span}{argument as for \code{voom}.}

\item{plot}{logical, as for \code{voom}.}

\item{save.plot}{logical, as for \code{voom}.}
}
\value{
An EList object as \code{voom}. E is numeric matrix of transformed counts on the arcsine scale
}
\description{
This function is called within \code{compute_stat()} and
takes argumetns as for \code{voom} in \url{https://bioconductor.org/packages/release/bioc/html/edgeR.html}
}
