% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcsinhlstLink.R
\name{arcsinhlstLink}
\alias{arcsinhlstLink}
\title{link function to be used in the negative binomial regression fitting}
\usage{
arcsinhlstLink()
}
\value{
A list of linkFun (link function), varFun(variance function),
InvLink (inverse of link function), and InvLinkDeriv (derivative of inverse of the link function)
}
\description{
link function to be used in the negative binomial regression fitting
}
