% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcsinhTransform.R
\name{arcsinhTransform}
\alias{arcsinhTransform}
\title{arcsinhTransform}
\usage{
arcsinhTransform(counts, design = NULL, lib.size = NULL, span = 0.5, ...)
}
\arguments{
\item{counts}{\code{DGEList}.}

\item{design}{a model matrix created by \code{computeStat()}.}

\item{lib.size}{numeric vector containing library size normalization factor computed by \code{computeStat()}.}

\item{span}{width of the lowess smoothing window as a proportion.}
}
\value{
An EList object. E is numeric matrix of transformed counts on the arcsine scale
}
\description{
This is a modified \code{voom} utility function used by the higher level function \code{computeStat()}
to compute arcsinh transformation and weights for the \code{glm.nb} and \code{geem}.
}
