% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboot_indices.R
\name{bboot_indices}
\alias{bboot_indices}
\title{bboot_indices}
\usage{
bboot_indices(x, b, time, k, blks_first_index)
}
\arguments{
\item{x}{Dataframe for each subject with repeated samples with \code{Index} variable inserted.}

\item{b}{Block lenght to account for dependence within-subject.}

\item{time}{Time variable at repeated measures.}

\item{k}{number of overlapping blocks per sample}

\item{blks_first_index}{randomly select first index of each block (same for all subjects)}
}
\value{
For each subject, this method returns indices of repeated samples for the block bootstrap sample as a list.
}
\description{
This function is called within \code{bboot_phyloseq} after
creating \code{Index} for samples within each subject
}
