% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboot_phyloseq.R
\name{bboot_phyloseq}
\alias{bboot_phyloseq}
\title{bboot_phyloseq}
\usage{
bboot_phyloseq(ps, b, time)
}
\arguments{
\item{ps}{a \code{phyloseq} object.}

\item{b}{a numeric. Block size.}

\item{time}{a character. Name of the time variable.}
}
\value{
a \code{phyloseq} object. This will be a block bootstrap realization of input \code{ps}.
}
\description{
Creates block bootstrap realizations as \code{phyloseq} objects.
}
