% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLongIndices.R
\name{bootLongIndices}
\alias{bootLongIndices}
\title{bootLongIndices}
\usage{
bootLongIndices(x, b, time, L, blks_first_index)
}
\arguments{
\item{x}{a data frame. It must have \code{Index} column that will be added by \code{bootLongPhyloseq}.}

\item{b}{a numeric. Block size.}

\item{time}{a character. Name of the time variable.}

\item{L}{a numeric. Number of overlapping blocks per subject that will be added by \code{bootLongPhyloseq}.}

\item{blks_first_index}{a vector of positive integer. This will be added by \code{bootLongPhyloseq} (same for all subjects.)}
}
\value{
A list of indices to include in the block bootstrap realization.
}
\description{
This function is called within \code{bootLongPhyloseq} for
each subject's data frame.
}
