% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLongMethod.R
\name{bootLongMethod}
\alias{bootLongMethod}
\title{bootLongMethod}
\usage{
bootLongMethod(ps, main_factor, time_var, subjectID_var, b, R, RR,
  FDR = 0.1)
}
\arguments{
\item{ps}{phyloseq object}

\item{main_factor}{Character string. The name of the variable to map main factor in the plot (factor vector).}

\item{time_var}{Character string. The name of the variable to map time in the plot (integer vector).}

\item{subjectID_var}{Character string. The name of the variable to map Subject Ids in the plot (factor vector).}

\item{b}{A numeric. The block size to account for the dependence within-subject.}

\item{R}{A numeric. The number of block bootstrap realizations.}

\item{RR}{A numeric. The number of double block bootstrap realizations.}

\item{FDR}{False discovery rate.}
}
\value{
list of dataframe with ASV, observed stat, pvalues, adjusted pvalues, lcl, ucl, observed pivotal quantity; stat.obs; stat.star; stat.star.star; T.star_obs.
}
\description{
bootLongMethod
}
