% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLongMethod.R
\name{bootLongMethod}
\alias{bootLongMethod}
\title{bootLongMethod}
\usage{
bootLongMethod(ps, b, R, RR, factors, time, FDR = 0.1)
}
\arguments{
\item{ps}{Observed \code{phyloseq} class object.}

\item{b}{numeric, optimal block size to account for dependence within-subject.}

\item{R}{Number of block bootstrap realization.}

\item{RR}{Number of double block bootstrap realization.}

\item{factors}{vector of factor variable(s) in the sample data of ps.}

\item{time}{Time variable at repeated observations.}

\item{FDR}{False discovery rate}
}
\value{
list of dataframe with ASV, observed stat, pvalues, adjusted pvalues, lcl, ucl, observed pivotal quantity; stat.obs; stat.star; stat.star.star; T.star_obs.
}
\description{
bootLongMethod
}
