% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLongMSEPsi.R
\name{bootLongMSEPsi}
\alias{bootLongMSEPsi}
\title{bootLongMSEPsi}
\usage{
bootLongMSEPsi(ps, main_factor, time_var, subjectID_var, sampleID_var, b,
  R, RR, qj, Wj, Khat.obs = NULL, T.obs.full = NULL, ncores)
}
\arguments{
\item{ps}{phyloseq object.}

\item{main_factor}{Character string. The name of the covariate variable.}

\item{time_var}{Character string. The name of the time variable.}

\item{subjectID_var}{Character string. The name of the subject ID variable.}

\item{sampleID_var}{Character string. The name of the sample ID variable.}

\item{b}{A numeric. The block size to account for the dependence within-subject.}

\item{R}{A numeric. The number of block bootstrap realizations.}

\item{RR}{A numeric. The number of double block bootstrap realizations.}

\item{qj}{A numeric vector. The number of repeated observations for j-th subject.}

\item{Wj}{A numeric vector. The number of repeated observations in the subset for j-th subject}

\item{Khat.obs}{A numeric vector. The second element of the output of \code{bootLongPsi} evaluated using the initial block size and full data.}

\item{T.obs.full}{A numeric vector. The numeric vector of observed statistic.}

\item{ncores}{A numeric. The number of cores to use in \code{\link[parallel]{mclapply}}.}
}
\value{
A list of MSE computed with given block size b, Khat with all subsamples, Khat with initial block size.
}
\description{
Compute MSE in computing \eqn{\psi} = two-sided probability with different block sizes.
}
