% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLongPhyloseq.R
\name{bootLongPhyloseq}
\alias{bootLongPhyloseq}
\title{bootLongPhyloseq}
\usage{
bootLongPhyloseq(ps, b, time, SubjectID_n = "SubjectID")
}
\arguments{
\item{ps}{a \code{phyloseq} object.}

\item{b}{a numeric. Block size.}

\item{time}{a character. Name of the time variable.}
}
\value{
a \code{phyloseq} object. This will be a block bootstrap realization of input \code{ps}.
}
\description{
Creates block bootstrap realizations as \code{phyloseq} objects.
}
