% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLongPhyloseq.R
\name{bootLongPhyloseq}
\alias{bootLongPhyloseq}
\title{bootLongPhyloseq}
\usage{
bootLongPhyloseq(ps, time_var, subjectID_var, sampleID_var, b)
}
\arguments{
\item{ps}{phyloseq object}

\item{time_var}{Character string. The name of the variable to map time in the plot (integer vector).}

\item{subjectID_var}{Character string. The name of the variable to map Subject Ids in the plot (factor vector).}

\item{b}{A numeric. The block size to account for the dependence within-subject.}
}
\value{
a \code{phyloseq} object. A block bootstrap realization of \code{ps}.
}
\description{
Creates block bootstrap realization of a \code{phyloseq}.
}
