% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLongPsi.R
\name{bootLongPsi}
\alias{bootLongPsi}
\title{bootLongPsi}
\usage{
bootLongPsi(ps, b, R, RR, factors, time, T.obs.full = NULL,
  SubjectID_n = "SubjectID", para = FALSE)
}
\arguments{
\item{ps}{Observed \code{phyloseq} class object.}

\item{b}{numeric, block size to account for dependence within-subject.}

\item{R}{Number of block bootstrap realization.}

\item{RR}{Number of double block bootstrap realization.}

\item{factors}{vector of factor variable(s) in the sample data of ps.}

\item{time}{Time variable at repeated observations.}

\item{T.obs.full}{If observed statistic is already computed}
}
\value{
a list with first element ``K.val`` - two-sided significance probability.
        second element ``observed statistic``
}
\description{
Compute \eqn{\psi} = two-sided probability with a given block size
}
