% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLongPsi.R
\name{bootLongPsi}
\alias{bootLongPsi}
\title{bootLongPsi}
\usage{
bootLongPsi(ps, main_factor = main_factor, time_var = time_var,
  subjectID_var = subjectID_var, sampleID_var = sampleID_var, b, R, RR,
  T.obs.full = NULL, ncores)
}
\arguments{
\item{ps}{phyloseq object.}

\item{main_factor}{Character string. The name of the covariate variable.}

\item{time_var}{Character string. The name of the time variable.}

\item{subjectID_var}{Character string. The name of the subject ID variable.}

\item{sampleID_var}{Character string. The name of the sample ID variable.}

\item{b}{A numeric. The block size to account for the dependence within-subject.}

\item{R}{A numeric. The number of block bootstrap realizations.}

\item{RR}{A numeric. The number of double block bootstrap realizations.}

\item{T.obs.full}{A numeric vector. The numeric vector of observed statistic.}

\item{ncores}{A numeric. The number of cores to use in \code{\link[parallel]{mclapply}}.}
}
\value{
a list with first element ``K.val`` - two-sided significance probability.
        second element ``observed statistic``
}
\description{
Compute \eqn{\psi} = two-sided probability with a given block size.
}
