% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLongSubsampling.R
\name{bootLongSubsampling}
\alias{bootLongSubsampling}
\title{bootLongSubsampling}
\usage{
bootLongSubsampling(ps, main_factor, time_var, subjectID_var, sampleID_var,
  lI, R, RR, omega = 0.6, lC1 = 1, lC2 = NULL, ncores = 3)
}
\arguments{
\item{ps}{phyloseq object}

\item{main_factor}{Character string. The name of the variable to map main factor in the plot (factor vector).}

\item{time_var}{Character string. The name of the variable to map time in the plot (integer vector).}

\item{subjectID_var}{Character string. The name of the variable to map Subject Ids in the plot (factor vector).}

\item{lI}{A numeric. The initial block size.}

\item{R}{A numeric. The number of block bootstrap realizations.}

\item{RR}{A numeric. The number of double block bootstrap realizations.}

\item{omega}{A numeric. The proportion between 0 and 1 of repeated observations for subsampling.}

\item{lC1}{A numeric. (optional) any block size to start with. defualt is 1.}

\item{lC2}{A numeric. (optional) any block size to end with. defualt is (lI-1).}

\item{ncores}{A numeric. The number of cores to use in \code{\link[parallel]{mclapply}}.}
}
\value{
A list of ``MSE`` in calculating K with block sizes lC=1:(lI-1), ``Khat`` values with block sizes lC=1:(lI-1) and ``Khat.obs`` with lI
}
\description{
Finding optimal block size
}
