% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLongSubsampling.R
\name{bootLongSubsampling}
\alias{bootLongSubsampling}
\title{bootLongSubsampling}
\usage{
bootLongSubsampling(ps, R, RR, factors, time, SubjectID_n = "SubjectID",
  lI, omega = 0.6, lC1 = 1, lC2 = NULL)
}
\arguments{
\item{ps}{a \code{phyloseq} object.}

\item{R}{numeric. Number of block bootstrap realizations.}

\item{RR}{numeric. Number of double block bootstrap realization.}

\item{factors}{vector of factor variable(s) in the sample data of ps.}

\item{time}{Time variable at repeated observations.}

\item{lI}{initial block size}

\item{omega}{percentage of repeated observations for subsampling}

\item{subjectidvar}{variable name for SubjectID}
}
\value{
list of ``MSE`` in calculating K with block sizes lC=1:(lI-1), ``Khat`` values with block sizes lC=1:(lI-1) and ``Khat.obs`` with lI
}
\description{
Finding optimal block size
}
