% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStat.R
\name{computeStat}
\alias{computeStat}
\title{computeStat}
\usage{
computeStat(ps, factors, time, b)
}
\arguments{
\item{ps}{Observed \code{phyloseq} class object.}

\item{factors}{vector of factor variable(s) in the sample data of ps.}

\item{time}{Time variable at repeated observations.}

\item{b}{numeric, optimal block size to account for dependence within-subject.}
}
\value{
dataframe with the last column corresponds to taxa names
and other columns are for estimated coefficients for each factors using GEE, including Intercept.
}
\description{
This function computes size factors for each samples using \code{estimateSizeFactors}.
Then, it transforms count data to inverse hyperbolic sine transformation.
Using the transformed counts, this function estimates the mean-variance relationship and
appropriate observation-level weights. These weights and the transformed data is used to fit a linear model
for each taxa. The empirical Bayes is applied to compute the shirnkage estimates for each factor.
}
