% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longCorreloMultiple.R
\name{longCorreloMultiple}
\alias{longCorreloMultiple}
\title{longCorreloMultiple}
\usage{
longCorreloMultiple(ps, factors, time, starttaxa = 1, endtaxa = 4,
  taxlevel = "Species")
}
\arguments{
\item{ps}{\code{phyloseq} object}

\item{factors}{factor, variable in the sample data of ps.}

\item{time}{character, time variable at repeated observations.}

\item{starttaxa}{numeric, first index of taxa ordered by total reads}

\item{endtaxa}{numeric, the last index of taxa ordered by total reads}

\item{taxlevel}{character, taxonomy level to put the title}
}
\value{
\code{ggplot2} object of correlogram for multiple taxa.
}
\description{
This function plots the correlogram given the taxa indices.
}
