% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longCorreloSingle.R
\name{longCorreloSingle}
\alias{longCorreloSingle}
\title{longCorreloSingle}
\usage{
longCorreloSingle(ps, factors, time, taxon, taxlevel = "Species")
}
\arguments{
\item{ps}{\code{phyloseq} object}

\item{factors}{factor variable in the sample data of ps.}

\item{time}{character, time variable at repeated observations.}

\item{taxon}{numeric, index of taxon to get the variogram}

\item{taxlevel}{character, taxonomy level to put the title}
}
\value{
\code{ggplot2} object of correlogram for one taxon according to the taxon index.
}
\description{
This function plots the correlogram given the taxon index.
}
