% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longLagPlot.R
\name{longLagPlot}
\alias{longLagPlot}
\title{Plots the lagged-plot given the taxon index.}
\usage{
longLagPlot(ps, main_factor, time_var, taxon, lags, taxlevel = "Species")
}
\arguments{
\item{ps}{(Required).  A \code{\link{phyloseq-class}}.}

\item{main_factor}{Character string. The name of the variable to map main factor in the plot (factor vector).}

\item{time_var}{Character string. The name of the variable to map time in the plot (integer vector).}

\item{taxon}{A numeric. The index of taxon to get the variogram}

\item{lags}{numeric (vector) how many lags to plot}

\item{taxlevel}{Character string. The taxonomy level for the plot title}
}
\value{
\code{ggplot2} object of lagged plot for speciec taxon.
}
\description{
Plots the lagged-plot given the taxon index.
}
