% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongVarioMultiple.R
\name{longVarioMultiple}
\alias{longVarioMultiple}
\title{Plots the variogram for multiple taxa given the indices.}
\usage{
longVarioMultiple(pstr, psres, main_factor, time_var, subjectID_var,
  starttaxa = 1, endtaxa = 4, point = FALSE, taxlevel = "Species")
}
\arguments{
\item{pstr}{phyloseq object. The asinh transfomed of the otu_table.}

\item{psres}{phyloseq object. The asinh transformed of the response residuals.}

\item{main_factor}{Character string. The name of the variable to map main factor in the plot (factor vector).}

\item{time_var}{Character string. The name of the variable to map time in the plot (integer vector).}

\item{subjectID_var}{Character string. The name of the variable to map Subject Ids in the plot (factor vector).}

\item{starttaxa}{Numeric. The first index of taxa ordered by total reads}

\item{endtaxa}{Numeric. The last index of taxa ordered by total reads}

\item{point}{Logical. Whether variogram with observed values.}

\item{taxlevel}{Character string. The taxonomy level for the plot title}
}
\value{
\code{ggplot} object of variogram for for multiple taxa.
}
\description{
Plots the variogram for multiple taxa given the indices.
}
