% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongVarioSingle.R
\name{longVarioSingle}
\alias{longVarioSingle}
\title{longVarioSingle}
\usage{
longVarioSingle(ps, factors, time, taxon, point = FALSE,
  taxlevel = "Species")
}
\arguments{
\item{ps}{\code{phyloseq} object}

\item{factors}{factor, variable in the sample data of ps.}

\item{time}{character, time variable at repeated observations.}

\item{taxon}{numeric, index of taxon to get the variogram}

\item{taxlevel}{character, taxonomy level to put the title}
}
\value{
\code{ggplot} object of variogram
}
\description{
This function plot the variogram given the tax index.
}
