% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psStanfordA.R
\docType{data}
\name{psStanfordA}
\alias{psStanfordA}
\title{The Stanford pregnancy data \code{'phyloseq'} object.}
\format{An object of class \code{'phyloseq'}.
\describe{
  \item{SubjectID}{IDs of individuals}
  \item{Preterm}{A factor variable with two levels TRUE or FALSE}
  \item{SampleID}{Sample ID for each repeated biological sample}
  \item{GWColl}{Gestational week at sampling}
}}
\source{
The data use in the paper: \href{https://www.ncbi.nlm.nih.gov/pubmed/26283357}{where}
}
\usage{
psStanfordA
}
\description{
A \code{'phyloseq'} class object with 109 taxa from 678 biological samples.
}
\examples{
\dontrun{sample_data(psStanfordA)
otu_table(psStanfordA)}
}
\references{
\href{https://www.ncbi.nlm.nih.gov/pubmed/26283357}{Study}
}
\keyword{datasets}
