% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psTransform.R
\name{psTransform}
\alias{psTransform}
\title{psTransform}
\usage{
psTransform(ps, main_factor)
}
\arguments{
\item{ps}{phyloseq object.}

\item{main_factor}{Character string. The name of the variable to map main factor in the plot (factor vector).}
}
\value{
A list of phyloseq objects. The first phyloseq element contains the arcsinh transformed of otu_table and the second phyloseq element contains the residuals of \code{\link[MASS]{glm.nb}} fit in otu table.
}
\description{
Variance-stablized values of otu table and residuals of glm.nb fit.
}
