% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psTransform.R
\name{psTransform}
\alias{psTransform}
\title{psTransform
This function takes the \code{phyloseq} object and apply
the arcsinh transformation.}
\usage{
psTransform(ps, factors)
}
\arguments{
\item{ps}{\code{phyloseq} object}

\item{factors}{factor, variable in the sample data of ps.}
}
\value{
\code{phyloseq} object with transformed \code{otu_table}
}
\description{
psTransform
This function takes the \code{phyloseq} object and apply
the arcsinh transformation.
}
