% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psTransform.R
\name{psTransform}
\alias{psTransform}
\title{arcsinh transformation of counts}
\usage{
psTransform(ps, main_factor)
}
\arguments{
\item{ps}{(Required). A \code{\link{phyloseq-class}}.}

\item{main_factor}{Character string. The name of the variable to map main factor in the plot (factor vector).}
}
\value{
A \code{\link{phyloseq-class}} with arcsinh transformed of otu_table.
}
\description{
Variance-stabilizing transformation of otu table.
}
